\name{print.yai}
\alias{print.yai}
\alias{summary.yai}
\title{Print a summary of a yai object}

\description{Provides a summary of a \code{\link{yai}} object, showing the
   call and essential data customized for each method used.}

\usage{
\method{print}{yai}(x,\dots)
}

\arguments{
  \item{x}{an object of class \code{yai}.}
  \item{\dots}{not used}
}

\author{
  Nicholas L. Crookston \email{ncrookston.fs@gmail.com} \cr
  Andrew O. Finley \email{finleya@msu.edu}
}

\keyword{print}

